Resolved conflicts in 8 files:
- shared/models/Subscription/SubscriptionParsers.ts: kept HEAD's null→undefined
  handling (matches Subscription.types.ts) and HEAD's (obj) signature so id is
  preserved in the result.
- src/components/AppShell/AppShell.tsx: took develop's isChatPanelAvailable
  check (proper abstraction over currentWorkspace).
- src/components/ChatPanel/useAvandarChatRuntime.ts: kept HEAD's plain `model`
  (the `?? undefined` from develop was a no-op).
- supabase/functions/subscriptions/subscriptions.routes.ts: kept HEAD's
  hardened customer-portal handler (user authz + multi-row find).
- supabase/schemas/07.subscriptions.sql: kept develop's more detailed Polar
  field comments (columns are identical).
- tests/e2e/csv-parse-options.spec.ts: kept HEAD's e2e fixes (triple-click +
  fill + Tab; re-upload to reset delimiter).
- tests/e2e/save-to-dashboard-renders-in-editor.spec.ts: kept HEAD's
  PTRCK-022 flake hardening (LONG_WAIT, explicit dashboard rename, sidebar
  navigation, toPass retry around the iframe assertion); dropped the
  stray duplicate TARGET_DASHBOARD_NAME.
- tests/e2e/share-modal.spec.ts: adopted develop's consolidated naming
  (dropped v2 suffix); helpers from datasetSharingFlowV2.ts moved to
  shareModalFlow.ts in develop, so the v2 helper file was removed.

pnpm type-check and pnpm lint both pass on the merged tree.

+++ supabase/schemas/20.datasets__xlsx_file.sql
@@ -15,7 +15,7 @@ create table public.datasets__xlsx_file (
   -- finished uploading), we set this to false.
   is_in_cloud_storage boolean not null default false,
   -- Size of the spreadsheet file in bytes
-  size_in_bytes integer not null,
+  size_in_bytes bigint not null,
   -- Number of rows to skip at the start of the imported worksheet
   rows_to_skip integer not null default 0,
   -- Name of the worksheet that was read. Nullable when the default sheet was
